refactor(icons): migrate build scripts to typescript
refs #DS-2461
