+++ packages/icons/config/tsconfig.test.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "moduleResolution": "node",
+    "module": "es2015"
+  },
+  "include": ["../scripts/**/*", "../jest.config.ts"]
+}
